Update module github.com/hetznercloud/hcloud-go to v2 (#72)
* Update module github.com/hetznercloud/hcloud-go to v2
* Use int64s instead of ints for IDs related to hcloud-go

---------

+++ pkg/hcloud/apis/types_worker.go
@@ -33,8 +33,8 @@ type WorkerStatus struct {
 	// resources that are still using this version. Hence, it stores the used versions in the provider status to ensure
 	// reconciliation is possible.
 	// +optional
-	MachineImages     []MachineImage `json:"machineImages,omitempty"`
-	PlacementGroupIDs map[string]int `json:"placementGroupIds,omitempty"`
+	MachineImages     []MachineImage   `json:"machineImages,omitempty"`
+	PlacementGroupIDs map[string]int64 `json:"placementGroupIds,omitempty"`
 }
 
 // MachineImage is a mapping from logical names and versions to provider-specific machine image data.
